+++ scripts/game_structure/ui_elements.py
@@ -36,35 +36,6 @@ from scripts.events_module.text_adjust import shorten_text_to_fit
 from scripts.ui.scale import ui_scale, ui_scale_dimensions, ui_scale_value
 
 
-class UIModifiedHorizScrollBar(pygame_gui.elements.UIHorizontalScrollBar):
-    def __init__(
-        self,
-        relative_rect: RectLike,
-        visible_percentage: float,
-        manager,
-        container,
-        parent_element,
-        anchors,
-        visible,
-        starting_height=1,
-    ):
-        super().__init__(
-            relative_rect,
-            visible_percentage,
-            manager=manager,
-            container=container,
-            parent_element=parent_element,
-            anchors=anchors,
-            visible=visible,
-        )
-
-        self.button_width = ui_scale_value(15)
-        self.arrow_button_width = self.button_width
-        self.sliding_button.change_layer(starting_height + 1)
-
-        self.rebuild()
-
-
 class UIRelationStatusFillBar(pygame_gui.elements.UIStatusBar):
     """Wraps together a status bar"""
